requirements.txt
PyYAML>=6.0
Flask>=3.0.0
Flask-CORS>=4.0.0
tencentcloud-sdk-python>=3.0.0
